requirements.txt
selene==2.0.0rc9
selenium
pytest
allure-pytest
allure-python-commons~=2.14.0
